Remove Profile fields other than name
None of the fields except profile_name is ever used at present (we don't
even do any validation that the encoded videos added to VAL match the
profile), and the edX media team prefers that the profiles be somewhat
fungible anyway. The motivation for this change is to avoid the need to
sync profile information between systems when doing export/import of
video modules that use Video IDs.

JIRA: MA-110