DE-1824: step 1 of 4 in renaming 'exists' field in ThirdPartyTranscriptCredentialsState to 'has_creds'  (#191)
* DE-1824: Add has_creds field to replace exists in 
ThirdPartyTranscriptCredentialsState, to eventually replace the
'exists' column.
* Update version number.
* Leave the API method signature unchanged.

The 'exists' column name in ThirdPartyTranscriptCredentialsState is a reserved keyword in SQL, so this PR starts the process of renaming the field to 'has_creds' in its place. This is step 1, to add the field to replace it. Step 2 will be to use a migration to copy the contents of existing 'exists' to 'has_creds'. Step 3 will be to remove the logic that relies on 'exists', and step 4 will be to add a migration to remove the 'exists' column from the database. This will also require bumping the version of edx-val, and updating the other repo that use edx-val, i.e. edx-platform.