SE-3540 [LX-1653] Fix create_transcripts_xml when resource_fs is WrapFS (#267)
* Fix create_transcripts_xml when resource_fs is WrapFS

`resource_fs` can be set to a WrapFS instance in some cases,
notably in override_export_fs, which is used when exporting content via
the olx_rest_api.

This bug was introduced in https://github.com/edx/edx-val/pull/258,
which in turn fixed other bugs.
So this commit preserves the behaviour from that PR,
while ensuring that the original static_file_dir logic is used when
a WrapFS is passed (ie. no _sub_dir attr).

An alternate option would be to add AttributeError to an except clause,
but it feels safer to be more specific and proactive in this case.