Updates Tox and Travis CI testing and documents the changes
This changes removes the awareness of the travis ci's configuration about the different tox testing environments. This is done by making the tox configuration solely responsible for storing the different environments that should be run for a specific python/django combination.

In addition, the updated test command in the Makefile makes it easy to automate running the tox environments which are specific to the existent python version.

This also makes it easy for developers to run all important tox testing environments, which would already be specified there.

Not to mention, this makes it much simpler to update automated tests using the tox configuration, without exerting a lot of effort on the travis ci configuration, which only requires now the python version.